remove vsync_recorder
