+++ shell/platform/fuchsia/flutter/BUILD.gn
@@ -83,8 +83,6 @@ template("runner_sources") {
       "thread.cc",
       "thread.h",
       "unique_fdio_ns.h",
-      "vsync_recorder.cc",
-      "vsync_recorder.h",
       "vsync_waiter.cc",
       "vsync_waiter.h",
       "vulkan_surface.cc",
@@ -455,7 +453,6 @@ executable("flutter_runner_unittests") {
     "runner_unittest.cc",
     "tests/engine_unittests.cc",
     "tests/flutter_runner_product_configuration_unittests.cc",
-    "tests/vsync_recorder_unittests.cc",
   ]
 
   # This is needed for //third_party/googletest for linking zircon symbols.
