+++ shell/platform/fuchsia/flutter/vsync_waiter.cc
@@ -15,7 +15,6 @@
 #include "flutter/fml/trace_event.h"
 
 #include "flutter_runner_product_configuration.h"
-#include "vsync_recorder.h"
 
 namespace flutter_runner {
